remove specific policy from set
Remove a specific (by namespace and name) policy from the PolicyEngine state.
This does not preserve slide order (which may requirre copying many/all elements
but instead moves the last policy in the slice to the removed index and then
shortens the slice by one.
The commit disable's revive's "add-constant" linter on the code as it wants to
replace "len(policies)-1" with a meaningfull constant...